Change UoM of VAR to var and kVAh to kWh
My hass instance was showing errors like

```
... sensor.sma_sunny_home_manager_2_reactive_power_consumption ... is using
native unit of measurement 'VAR' which is not a valid unit for the device
class ('reactive_power') it is using; expected one of ['var'] ...
```

and

```
... sensor.sma_sunny_home_manager_2_reactive_power_consumption_counter ... is
using native unit of measurement 'kVAh' which is not a valid unit for the
device class ('energy') it is using; expected one of ['MWh', 'kWh', 'MJ',
'GJ', 'Wh'] ...
```

I suspect the hass core became a little stricter on these units since I think
this only started popping up after an upgrade.

Note that this is a breaking change. After the changes, I saw the following logs:

```
The unit of sensor.sma_sunny_home_manager_2_phase_1_apparent_power_consumption_counter
is changing, got multiple {'kVAh', 'kWh'}, generation of long term statistics
will be suppressed unless the unit is stable and matches the unit of already
compiled statistics (kVAh). Go to https://my.home-assistant.io/redirect/developer_statistics
to fix this
```